requirements.txt
flask
flask-cors
torch
torchvision
pillow
efficientnet_pytorch
gunicorn
pip>=25.2